Created the APIDetails Class and tested more functionality.
I created the APIDetails class to be close in line with the project checkpoint planning. I still need to figure out how/what way we will take in user input so that I can search the api for the user input song. Make sure to use a format like "song name artist name" to ensure we get the exact right song.