Update to Pull Lyrics from Genius through Python Wrapper.
Can now pull lyrics from songs via lyricsgenius python wrapper. Calling a python module from Java class. Need to make the .py file an executable and package it so we can call it from relative path and the user doesn't need to have python to call it. The lyrics are stored within the APIDetails object as a String field.